runmodes: allow more than 256 threads
Don't cast int to uint8_t for no reason. Add warning that upper
limit is 1024.

Small code cleanups.

Bug: #2228

+++ src/runmode-af-packet.c
@@ -188,7 +188,7 @@ static void *ParseAFPConfig(const char *iface)
             if (strcmp(threadsstr, "auto") == 0) {
                 aconf->threads = 0;
             } else {
-                aconf->threads = (uint8_t)atoi(threadsstr);
+                aconf->threads = atoi(threadsstr);
             }
         }
     }
